pyproject.toml
[tool.poetry]
name = "interview-assistant"
version = "0.1.0"
description = ""
authors = ["Carlo Giuseppe Sergi <c.sergi@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.64.0"
pyqt5 = "^5.15.11"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
pyperclip = "^1.9.0"
pyautogui = "^0.9.54"
requests = "^2.32.3"
pyaudio = "^0.2.14"
websocket-client = "^1.8.0"
numpy = "^2.2.3"
sounddevice = "^0.5.1"
soundfile = "^0.13.1"
textual = "^2.1.2"
pydub = "^0.25.1"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
